+++ chat-server/app/core/config.py
@@ -1,4 +1,3 @@
-import logging
 from typing import List, Union
 
 from pydantic import AnyHttpUrl, AnyUrl, BaseSettings, validator, AmqpDsn, BaseModel
