+++ master-server/app/models.py
@@ -1,0 +1,10 @@
+from sqlalchemy import Column, Integer, String, LargeBinary
+from app.database import Base
+
+
+class User(Base):
+    __tablename__ = "users"
+
+    id = Column(Integer, primary_key=True, index=True)
+    login = Column(String, unique=True, index=True)
+    password = Column(LargeBinary)
