+++ master-server/app/settings.py
@@ -6,6 +6,7 @@ class Settings(BaseSettings):
     JWT_SECRET: str
     DATABASE_URL: str
     BROKER_URL:str
+    HEARTBEAT_EXCHANGE_NAME: str = "heartbeat-exchange"
 
     class Config:
         env_file = ".env"
